Conflicts:
	fabtools/files.py
	fabtools/postgres.py
	fabtools/require/postgres.py
	fabtools/tests/test_postgres.py

+++ fabtools/tests/vagrant.py
@@ -1,14 +1,13 @@
 from __future__ import with_statement
 
 import os
-import os.path
 
 try:
     import unittest2 as unittest
 except ImportError:
     import unittest
 
-from fabric.api import *
+from fabric.api import env, hide, lcd, local, settings, shell_env
 from fabric.state import connections
 
 import fabtools
@@ -44,9 +43,9 @@ def base_boxes():
 
     The default is to get the list of all base boxes.
 
-    This can be overridden with the VAGRANT_BOXES environment variable.
+    This can be overridden with the FABTOOLS_TEST_BOXES environment variable.
     """
-    boxes = os.environ.get('VAGRANT_BOXES')
+    boxes = os.environ.get('FABTOOLS_TEST_BOXES')
     if boxes is not None:
         return boxes.split()
     else:
@@ -93,7 +92,8 @@ class VagrantTestSuite(unittest.BaseTestSuite):
 
             # Make sure the package index is up to date
             with self.settings():
-                fabtools.deb.update_index()
+                if fabtools.system.distrib_family() == 'debian':
+                    fabtools.deb.update_index()
 
             # Run the test suite
             unittest.BaseTestSuite.run(self, result)
@@ -105,18 +105,35 @@ class VagrantTestSuite(unittest.BaseTestSuite):
         """
         Spin up a new vagrant box
         """
+
+        # Support for Vagrant 1.1 providers
+        if ':' in self.current_box:
+            box_name, provider = self.current_box.split(':', 1)
+        else:
+            box_name = self.current_box
+            provider = None
+
         with lcd(os.path.dirname(__file__)):
 
-            # Create a fresh vagrant config file
-            local('rm -f Vagrantfile')
-            local('vagrant init %s' % self.current_box)
+            if not os.path.exists('Vagrantfile') \
+               or not os.environ.get('FABTOOLS_TEST_NODESTROY'):
 
-            # Clean up
-            halt_and_destroy()
+                # Create a fresh vagrant config file
+                local('rm -f Vagrantfile')
+                local('vagrant init %s' % box_name)
+
+                # Clean up
+                halt_and_destroy()
+
+            if provider:
+                options = ' --provider %s' % provider
+            else:
+                options = ''
 
             # Spin up the box
             # (retry as it sometimes fails for no good reason)
-            local('vagrant up || vagrant up')
+            cmd = 'vagrant up%s' % options
+            local('%s || %s' % (cmd, cmd))
 
     def ssh_config(self):
         """
@@ -140,9 +157,10 @@ class VagrantTestSuite(unittest.BaseTestSuite):
         """
         Spin down the vagrant box
         """
-        halt_and_destroy()
-        with lcd(os.path.dirname(__file__)):
-            local('rm -f Vagrantfile')
+        if not os.environ.get('FABTOOLS_TEST_NODESTROY'):
+            halt_and_destroy()
+            with lcd(os.path.dirname(__file__)):
+                local('rm -f Vagrantfile')
         self.current_box = None
 
     def settings(self, *args, **kwargs):
@@ -157,7 +175,7 @@ class VagrantTestSuite(unittest.BaseTestSuite):
 
         kwargs['host_string'] = "%s@%s:%s" % (user, hostname, port)
         kwargs['user'] = user
-        kwargs['key_filename'] = config['IdentityFile']
+        kwargs['key_filename'] = config['IdentityFile'].strip('"')
         kwargs['disable_known_hosts'] = True
 
         return settings(*args, **kwargs)
@@ -178,7 +196,9 @@ class VagrantTestCase(unittest.TestCase):
         Run the test case within a Fabric context manager
         """
         with self._suite.settings():
-            unittest.TestCase.run(self, result)
+            http_proxy = os.environ.get('FABTOOLS_HTTP_PROXY', '')
+            with shell_env(http_proxy=http_proxy):
+                unittest.TestCase.run(self, result)
 
     def runTest(self):
         self._callable()
